requirements.txt
black==23.9.0
click==8.1.7
pytest==7.4.2
pytest-cov==4.1.0
pylint==2.17.5
nbval==0.10.0
python-dotenv==1.0.0


openai==1.3.7
uvicorn==0.24.0
fastapi==0.104.1
jinja2==3.1.2